converted float-type variables to double type

+++ src/ICohort_PnET.cs
@@ -13,42 +13,42 @@ namespace Landis.Extension.Succession.PnETForC
         int AGBiomass { get; }
         int TotalBiomass { get; }
         bool IsLeafOn { get; }
-        float MaxBiomass { get; }
-        float Fol { get; }
-        float MaxFolYear { get; }
-        float NSC { get; }
-        float DefoliationFrac { get; }
-        float LastWoodSenescence { get; }
-        float LastFolSenescence { get; }
-        float LastFRad { get; }
-        List<float> LastSeasonFRad { get; }
-        float adjFolBiomassFrac { get; }
-        float AdjHalfSat { get; }
-        float adjFolN { get; }
+        double MaxBiomass { get; }
+        double Fol { get; }
+        double MaxFolYear { get; }
+        double NSC { get; }
+        double DefoliationFrac { get; }
+        double LastWoodSenescence { get; }
+        double LastFolSenescence { get; }
+        double LastFRad { get; }
+        List<double> LastSeasonFRad { get; }
+        double adjFolBiomassFrac { get; }
+        double AdjHalfSat { get; }
+        double adjFolN { get; }
         int ColdKill { get; }
         byte Layer { get; }
-        float[] LAI { get; }
-        float LastLAI { get; }
-        float LastAGBio { get; }
-        float[] GrossPsn { get; }
-        float[] FoliarRespiration { get; }
-        float[] NetPsn { get; }
-        float[] MaintenanceRespiration { get; }
-        float[] Transpiration { get; }
-        float[] PotentialTranspiration { get; }
-        float[] FRad { get; }
-        float[] FWater { get; }
-        float[] SoilWaterContent { get; }
-        float[] PressHead { get; }
+        double[] LAI { get; }
+        double LastLAI { get; }
+        double LastAGBio { get; }
+        double[] GrossPsn { get; }
+        double[] FoliarRespiration { get; }
+        double[] NetPsn { get; }
+        double[] MaintenanceRespiration { get; }
+        double[] Transpiration { get; }
+        double[] PotentialTranspiration { get; }
+        double[] FRad { get; }
+        double[] FWater { get; }
+        double[] SoilWaterContent { get; }
+        double[] PressHead { get; }
         int[] NumPrecipEvents { get; }
-        float[] FOzone { get; }
-        float[] Interception { get; }
-        float[] AdjFolN { get; }
-        float[] AdjFolBiomassFrac { get; }
-        float[] CiModifier { get; }
-        float[] DelAmax { get; }
-        float BiomassLayerFrac { get; }
-        float CanopyLayerFrac { get; }
-        float CanopyGrowingSpace { get; }
+        double[] FOzone { get; }
+        double[] Interception { get; }
+        double[] AdjFolN { get; }
+        double[] AdjFolBiomassFrac { get; }
+        double[] CiModifier { get; }
+        double[] DelAmax { get; }
+        double BiomassLayerFrac { get; }
+        double CanopyLayerFrac { get; }
+        double CanopyGrowingSpace { get; }
     }
 }
