+++ src/ISiteCohorts_PnET.cs
@@ -9,26 +9,26 @@ namespace Landis.Extension.Succession.PnETForC
 {
     public interface ISiteCohorts : Library.UniversalCohorts.ISiteCohorts
     {
-        float[] NetPsn { get; }
-        float[] MaintResp { get; }
-        float[] GrossPsn { get; }
-        float[] FoliarRespiration { get; }
-        float[] AverageAlbedo { get; }
-        float[] ActiveLayerDepth { get; }
-        float[] FrostDepth { get; }
-        float[] MonthlyAvgSnowpack { get; }
-        float[] MonthlyAvgWater { get; }
-        float[] MonthlyAvgLAI { get; }
-        float[] MonthlyEvap { get; }
-        float[] MonthlyActualTrans { get; }
-        float[] MonthlyLeakage { get; }
-        float[] MonthlyInterception { get; }
-        float[] MonthlyRunoff { get; }
-        float[] MonthlyActualET { get; }
-        float[] MonthlyPotentialEvap { get; }
-        float[] MonthlyPotentialTrans { get; }
-        float CanopyLAImax { get; }
-        float SiteMossDepth { get; }
+        double[] NetPsn { get; }
+        double[] MaintResp { get; }
+        double[] GrossPsn { get; }
+        double[] FoliarRespiration { get; }
+        double[] AverageAlbedo { get; }
+        double[] ActiveLayerDepth { get; }
+        double[] FrostDepth { get; }
+        double[] MonthlyAvgSnowpack { get; }
+        double[] MonthlyAvgWater { get; }
+        double[] MonthlyAvgLAI { get; }
+        double[] MonthlyEvap { get; }
+        double[] MonthlyActualTrans { get; }
+        double[] MonthlyLeakage { get; }
+        double[] MonthlyInterception { get; }
+        double[] MonthlyRunoff { get; }
+        double[] MonthlyActualET { get; }
+        double[] MonthlyPotentialEvap { get; }
+        double[] MonthlyPotentialTrans { get; }
+        double CanopyLAImax { get; }
+        double SiteMossDepth { get; }
         int AverageAge { get; }
         Library.Parameters.Species.AuxParm<int> CohortCountPerSpecies { get; }
         Library.Parameters.Species.AuxParm<bool> SpeciesPresent { get; }
@@ -40,28 +40,28 @@ namespace Landis.Extension.Succession.PnETForC
         Library.Parameters.Species.AuxParm<int> BGBiomassPerSpecies { get; }
         Library.Parameters.Species.AuxParm<int> FoliageBiomassPerSpecies { get; }
         Library.Parameters.Species.AuxParm<int> NSCPerSpecies { get; }
-        Library.Parameters.Species.AuxParm<float> LAIPerSpecies { get; }
+        Library.Parameters.Species.AuxParm<double> LAIPerSpecies { get; }
         Library.Parameters.Species.AuxParm<int> WoodSenescencePerSpecies { get; }
         Library.Parameters.Species.AuxParm<int> FolSenescencePerSpecies { get; }
         Library.Parameters.Species.AuxParm<List<ushort>> CohortAges { get; }
-        float BiomassSum { get; }
-        float AGBiomassSum { get; }
-        float WoodBiomassSum { get; }
-        float WoodSenescenceSum { get; }
-        float FolSenescenceSum { get; }
+        double BiomassSum { get; }
+        double AGBiomassSum { get; }
+        double WoodBiomassSum { get; }
+        double WoodSenescenceSum { get; }
+        double FolSenescenceSum { get; }
         int CohortCount { get; }
-        float JulySubCanopyPAR { get; }
-        float MaxSubCanopyPAR { get; }
+        double JulySubCanopyPAR { get; }
+        double MaxSubCanopyPAR { get; }
         double LeafLitter { get; }
         double WoodDebris { get; }
         int AgeMax { get; }
-        float AvgSoilWaterContent { get; }
-        float BGBiomassSum { get; }
-        float FolBiomassSum { get; }
-        float NSCSum { get; }
-        float ActualETSum { get; } //mm
-        float NetPsnSum { get; }
-        float PotentialET { get; }
+        double AvgSoilWaterContent { get; }
+        double BGBiomassSum { get; }
+        double FolBiomassSum { get; }
+        double NSCSum { get; }
+        double ActualETSum { get; } //mm
+        double NetPsnSum { get; }
+        double PotentialET { get; }
 
         List<ISpecies> SpeciesEstByPlanting { get; set; }
         List<ISpecies> SpeciesEstBySerotiny { get; set; }
